requirements.txt
appdirs==1.4.4
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Django==3.0.4
djangorestframework==3.11.0
flake8==3.8.1
idna==2.9
importlib-metadata==1.6.0
mccabe==0.6.1
multidict==4.7.5
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2019.3
regex==2020.5.14
requests==2.23.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
yarl==1.4.2
zipp==3.1.0
psycopg2==2.8.5
celery==4.4.2
redis==3.5.2
